# Improve speed of creating NanoDates from strings