@JeffreySarnoff 2022-07-27
Yes and Thank you. Nowadays speedier parsing for simpler types is provided with [Parsers.jl,](https://github.com/JuliaData/Parsers.jl) -- which is already incorporated in Project.toml and using_ed in NanoDates.jl; the functions are not yet called within our srcfiles.
I intend for their Integer and their Date parsers to be as a subfloor supporting NanoDates' parsing wherever it makes sense.

from their README, the section on  Date/DateTime parsing
```
using Dates
a = Parsers.parse(Date, "2018-01-01")

# custom dateformat
b = Parsers.parse(Date, "01/20/2018",
           Parsers.Options(dateformat="mm/dd/yyyy"))
```

I will ask for JuliaData's guidance in extending their parsers for types exported by Dates.jl and paste the link to that thread here.