@JeffreySarnoff 2022-07-27
For our purposes,  there are a few issues with the approach[es] Dates.jl takes -- I never saw any problem until working to clear the last of some parsing test cases [given by others from their actual needs].  After running into strong algorithmic stop signs I had to decide (a) copy a number of large sections from the Dates.jl source files and reprogram away the corner cases or (b) toss it in favor of a different approach. 
(a) was disheartening and hidden piracy (b) is where "make it work, then make if fast." applies.

On the other side of the mirror, where dates-with-times are pushed out as the same, now quite slowly, parsed strings, this happens and is made faster with attention, re-simplification and benchmarking.
The general dateformat strings that are intended to provide a soft and consistent place for temporal measures to land and reside,, transparently are designed from the DatePeriod and TimePeriod types and the substructural algebra that applies to some constituents. ---

NanoDates sees "yyyy-mm-ddTHH:MM:SS.ssssssZ" like this
                           "      -       -    T     :      :     .          Z"       with adjacent spaces as typed digit hammocks
This is very much more flexible and can be processed with periods handled in parallel.