@anowacki 2022-07-27
Okay, thanks for explaining and good to hear there is a Parsers.jl-backed implementation planned which I guess would be somewhat quicker.  I think I had misunderstood the approach based on my incorrect reading of #1.