@anowacki 2022-07-28
Okay, glad to have this short-term speedup for now.

I just looked at Parsers.jl.  The `DateTime` parsing is only about twice as slow as Dates's own when using the standard format, and only three times worse with a custom `DateFormat`, though it does allocate:

```julia
julia> using Dates, Parsers, BenchmarkTools

julia> @benchmark parse(DateTime, "2000-01-01T00:00:00.001")
BenchmarkTools.Trial: 10000 samples with 973 evaluations.
 Range (min … max):  73.181 ns … 153.527 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     77.379 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   80.134 ns ±   7.609 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▄▅▆▆▆█▆▇▅▅▄▃▃▁▂▁▁▁▁▁ ▂▂▂▂▃▁▂▃▁▁▂                             ▂
  ██████████████████████████████████▇█▇▇█▇█▇█▇█████▇▇▃▃▄▅▅▅▅▅▆ █
  73.2 ns       Histogram: log(frequency) by time       111 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark Parsers.parse(DateTime, "2000-01-01T00:00:00.001")
BenchmarkTools.Trial: 10000 samples with 896 evaluations.
 Range (min … max):  124.990 ns … 252.307 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     128.374 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   132.139 ns ±   9.320 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▃█▇▇▆▁▄▄▂▁ ▁▁        ▆▃ ▁ ▂      ▁                           ▂
  ▂██████████▅████▅▄██▆▇██▇████▇▇████▇▆▆▇▆▇▇▆▅▆▆▆▇▇▆▅▄▇▇▇▆▇▆▆▆▆ █
  125 ns        Histogram: log(frequency) by time        169 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark parse(DateTime, "2000-01-01T00:00:00.001", dateformat"yyyy-mm-ddTHH:MM:SS.sss")
BenchmarkTools.Trial: 10000 samples with 905 evaluations.
 Range (min … max):  120.779 ns … 244.760 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     127.523 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   132.245 ns ±  12.301 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▇█  ▄▅ ▇█▂▁▃▅ ▃▁▁▂ ▃▂▂▃ ▄▅▂▃ ▁  ▃▅         ▁▁                 ▂
  ██▆██████████▇████▅████████████▇█████▇██████████▇▇█▇▇▆▆▇▆▇▇▇█ █
  121 ns        Histogram: log(frequency) by time        172 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark Parsers.parse(DateTime, "2000-01-01T00:00:00.001", Parsers.Options(dateformat=dateformat"yyyy-mm-ddTHH:MM:SS.sss"))
BenchmarkTools.Trial: 10000 samples with 206 evaluations.
 Range (min … max):  355.388 ns …  15.810 μs  ┊ GC (min … max):  0.00% … 97.41%
 Time  (median):     362.529 ns               ┊ GC (median):     0.00%
 Time  (mean ± σ):   409.231 ns ± 779.923 ns  ┊ GC (mean ± σ):  10.08% ±  5.15%

  ▁▅▇██▆▅▄▃▃▃▂▁▁▂▂▃▂▁▁▂▁▁                                       ▂
  █████████████████████████▇▆▆▆▄▆▃▅▃▁▅▄▄▅▅▅▅▆▇▇▇██▇▇▇▇▆▆▆▅▅▆▄▅▅ █
  355 ns        Histogram: log(frequency) by time        461 ns <

 Memory estimate: 624 bytes, allocs estimate: 16.
```

Thought I would post that since I had run the benchmarks but you are probably well aware already.